requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.4
requests==2.28.1
beautifulsoup4==4.11.2
spacy==3.5.2
python-dotenv==1.0.0
